requirements.txt
sentence-transformers==4.1.0
langchain==0.3.26
flask==3.1.1
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-community==0.3.26
crewai==0.22.5
langgraph==0.1.19
crewai[tools]
-e .